Added boilerplate and infrastructure
This commit adds a lot of boilerplate code establishing program
structure and setting up the environment. It adds a gulpfile which
can be used to compile everything to a form that's browser-compatible.
Gulp is chosen arbitrarily. Note also that the gulp watch task and
sourcemaps do not yet work as desired.