Connection class, skeleton Channel, SignalingChannel, Circuit classes
This commit adds most of the boilerplate WebRTC code for establishing
peer connections. It doesn't yet implement the actual data channels,
however, and assumes a certain interface to signaling channels, though
these are also not yet fully defined. We have also added the skeleton of
a signaling channel based on the WebRTC Data channel, as described in
the design document. This will be the primary (non-bootstrap) signaling
channel. We have chosen to implement these abstractions as classes, as
this will allow for clearer abstraction boundaries and better data
management as we start to deal with many open connections, channels, and
circuits at once.