More signaling, bridge/server framework, message specs
This commit primarily adds more flesh to the messaging system, including
introducing the event-based architecture for the signaling channels. Now
anything that would like can subscribe to a signaling channel to run a
callback when messages of a particular type corresponding to a
particular id are received. A similar system will be used for the raw
Channels, as well, but is not yet implemented. This improved the
encapsulation behavior and reduces metadata storage as compared to the
previous architecture.