Slightly change network.js spec, add example comments
This commit adds some example usage snippets to the file comments for
network.js and signal.js corresponding to possible use cases. It also
changes the spec for Channel slightly to make it more convenient to use,
and adds the getChannel function to serve as the primary transparent
interface to the network layer.