Implement getChannel, add enc/dec for channel
This commit adds utilities in the Channel class to add stream
encryption. This is intended for encrypting the raw messages sent on the
channel using a user-defined encrytion mechanism. Payload or other
sublevel encryption can be done prior to direct sending over the wire.
This is primarily for the purposes of symmetric channel encryption
following a Diffie-Hellman key exchange on the unencrypted channel.