Added mocha, chai testing framework, gulp target, encoding
This commit adds the lowest-level encoding and decoding functions for
messages using the node Buffer polyfill for browsers so that node
buffers can be used both in node and in the browser (instead of
browser-only Blobs). It also adds a testing framework and the 'gulp
test' task for running tests. Test files should be put in the
`app/scripts/test` file and follow the mocha/chai format as seen in the
other files there and through examples online.