Encode/decode framework for remaining functions
This commit adds encoding and decoding framework for the remaining
top-level message types, as well as the payloads of RELAY and SIGNAL
messages, with almost all functions currently not implemented.
Implementations for these parts should take the form of the existing
types, primarily doing object validity verification and converting
between Buffers and JS object formats.