Add length to message header, socket signaling maybe working
This commit adds a length field to the message header so that packets
can be sent over raw sockets. It also adds a form element for connecting
to a bridge, and makes the bridge successfully receive packets. The
bridge cannot make a webrtc connection, however, as the wrtc library
isn't working, so this will need to be changed.